+++ client/src/js/app.js
@@ -1,16 +1,24 @@
 import Vue from "vue";
 import Axios from 'axios';
-import trail from './components/Trail.vue'
+import trial from './components/Trial.vue'
 import button from './components/Button.vue'
 
 
 Vue.prototype.$http = Axios;
-Vue.component('trail',trail);
-Vue.component('Button',button);
+Vue.component('trial',trial);
+Vue.component('xbutton',button);
 
 window.app = new Vue({ el:'#app',data: function(){
         return {
+            helloCount:0
+
         }
-}});
+},
+    methods:{
+        increaseHelloCount : function(ev) {
+            this.helloCount++;
+        }
+    },
+});
